Faltan cositas.

+++ Unidad1/Ejercicio03LecturaFicherosTexto/Product.java
@@ -6,7 +6,7 @@ public class Product implements Comparable<Product> {
     private double unitPrice;
     private int unitsInStock;
 
-    public Product(int id, String name, int supplier, int category, int unitPrice, int unitsInStock) {
+    public Product(int id, String name, int supplier, int category, double unitPrice, int unitsInStock) {
         this.id = id;
         this.name = name;
         this.supplier = supplier;
